Suggest the image-map package in the GraphViz composer.json.
The GraphViz extension depends on the ImageMap extension being
installed.  Since the ImageMap extension is bundled with MediaWiki,
the image-map package is merely a suggestion.

Also, since the image-map package only exists @dev stability,
the graph-viz package cannot unilaterally require/install it without
a change to minimum-stability in the main MediaWiki composer.json.

While the image-map composer package may be convenient for some,
it will cause the latest development version of the extension to
be installed with each "composer update" command.  For stability,
the MediaWiki admin may prefer to keep the snapshot of the ImageMap
extension that was bundled with the installed MediaWiki version.

Change-Id: I980892434bfaf345f05a37cb6badca439c55a2ca