Add tag argument “preparse”.
This change adds support for wiki variable and template
replacement in the wikitext before it is passed to the dot
or mscgen renderer.

The new tag argument “preparse” indicates that the wikitext
should be preparsed using Parser::recursiveTagParse.
The key/value pair preparse=”dynamic” means that the
preparse should occur every time the page is displayed.
The key/value pair preparse=”static” means that the
preparsing should occur only when an edit is saved or previewed.

Change-Id: I0cb1457ce9d04467e69d6992aed9655ad7d37606