Eliminate PHP fread() length 0 warning
The full text of the warning is:
PHP Warning: fread(): Length parameter must be
greater than 0 in /w/includes/MimeMagic.php on line 595

The copy from $imageFilePath to $dummyFilePath failed due to a
file that already existed at the destination $dummyFilePath
(despite PHP documentation saying copy will overwrite the
target if it exists).

The fix is to unlink the target if it exists before doing the copy.

Bug: T89325

Change-Id: I70ae9d8b8f0cad28fd741d8260efc50b7a2b4e10