Prevent double-encoding of URLs that were output by the 'dot' command
The output of 'dot -Tcmapx' encodes HTML entities in URLs, and because
this is passed straight to ImageMap::render (which then also encodes
these entities) the final output is double-encoded.

This change fixes this for non-mscgen renderers (i.e. only GraphViz,
because 'mscgen -T ismap' does not encode HTML entities in URLs).

Bug: T151294
Change-Id: I31851bfd3a08420eac8b33efd303fb6520370f30